rename file, responsive notification page
