+++ src/components/Account.jsx
@@ -47,7 +47,7 @@ const Account = () => {
             className="flex items-center bg-customBlue2 text-white px-4 py-3 rounded-xl w-full md:w-9/12 ml-2 md:ml-28 lg:ml-36 xl:ml-48 mt-8 hover:bg-customBlue1"
           >
             <FiArrowLeft size={24} className="mr-2" />
-            Back to Homepage
+            Beranda
           </Link>
         </div>
       </div>
@@ -142,7 +142,7 @@ const Account = () => {
                 type="submit"
                 className="bg-customBlue2 text-white px-10 py-3 rounded-xl self-center hover:bg-customBlue1"
               >
-                Save
+                Simpan
               </button>
             </form>
           </div>
