--- src/pages/Notification.jsx
@@ -1,13 +1,0 @@
-import Navbar from "../components/Navbar";
-import NotificationPage from "../components/NotificationPage";
-
-const Notification = () => {
-  return (
-    <>
-      <Navbar />
-      <NotificationPage />
-    </>
-  );
-};
-
-export default Notification;
