+++ src/pages/NotificationPage.jsx
@@ -1,0 +1,13 @@
+import Navbar from "../components/NavbarAccount";
+import Notification from "../components/Notification";
+
+const NotificationPage = () => {
+  return (
+    <>
+      <Navbar />
+      <Notification />
+    </>
+  );
+};
+
+export default NotificationPage;
